requirements.txt
Django>=3.2.1,<3.3
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
djangorestframework>=3.13.1,<3.14
django-cors-headers>=3.13.0,<3.14
moto[s3]
boto3>=1.26.66,<1.27
uwsgi>=2.0.20,<2.1